# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Rythmx API
  version: 1.0.0
  description: >
    Self-hosted music automation and library management API.
    All endpoints under `/api/v1/` follow a standard envelope:
    success responses include `"status": "ok"` plus payload fields;
    error responses include `"status": "error"` and `"message"`.
    Background-triggering endpoints return HTTP 202 with `"is_alive": bool`.
servers:
  - url: /
    description: Rythmx server (all paths are absolute)

# All endpoints require an API key (X-Api-Key header) except /health
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/stats/top-artists:
    get:
      tags: [stats]
      summary: Get top artists from Last.fm
      operationId: getTopArtists
      parameters:
        - $ref: '#/components/parameters/Period'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Top artists
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      artists:
                        type: array
                        items:
                          $ref: '#/components/schemas/Artist'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/stats/top-tracks:
    get:
      tags: [stats]
      summary: Get top tracks from Last.fm
      operationId: getTopTracks
      parameters:
        - $ref: '#/components/parameters/Period'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Top tracks
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      tracks:
                        type: array
                        items:
                          $ref: '#/components/schemas/Track'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/stats/top-albums:
    get:
      tags: [stats]
      summary: Get top albums from Last.fm
      operationId: getTopAlbums
      parameters:
        - $ref: '#/components/parameters/Period'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Top albums
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      albums:
                        type: array
                        items:
                          $ref: '#/components/schemas/TopAlbum'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/stats/summary:
    get:
      tags: [stats]
      summary: Get listening history summary
      operationId: getStatsSummary
      responses:
        '200':
          description: Stats summary
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - $ref: '#/components/schemas/StatsSummary'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/stats/loved-artists:
    get:
      tags: [stats]
      summary: Get loved artist names from Last.fm
      operationId: getLovedArtists
      responses:
        '200':
          description: Loved artist names
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      artists:
                        type: array
                        items:
                          type: string
        '500':
          $ref: '#/components/responses/ServerError'

  # ── Settings ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: >
        Auto-generated API key. View or regenerate in Settings.
        Included as `X-Api-Key: <key>` on every request.
  parameters:
...
    Period:
      name: period
      in: query
      schema:
        $ref: '#/components/schemas/Period'
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
  responses:
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    Period:
      type: string
      enum: [7day, 1month, 3month, 6month, 12month, overall]
...
    ApiSuccess:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
    ApiError:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [error]
        message:
          type: string
        error:
          type: string

    # ── Music Entities ───────────────────────────────────────────────────────
    Artist:
      type: object
      required: [name]
      properties:
        name:
          type: string
        image:
          type: string
        playcount:
          type: integer
        url:
          type: string
    TopAlbum:
      type: object
      required: [artist, title]
      properties:
        artist:
          type: string
        title:
          type: string
        image:
          type: string
        playcount:
          type: integer
    Track:
      type: object
      required: [name, artist]
      properties:
        name:
          type: string
        artist:
          type: string
        album:
          type: string
        image:
          type: string
        playcount:
          type: integer
        duration:
          type: integer
          description: Duration in seconds
        is_owned:
          type: boolean
        score:
          type: number
          format: float
        rank:
          type: integer

    # ── Acquisition ──────────────────────────────────────────────────────────
...
    StatsSummary:
      type: object
      required: [top_artist, total_artists]
      properties:
        top_artist:
          type: string
        total_artists:
          type: integer
        total_plays:
          type: integer

    # ── Settings ─────────────────────────────────────────────────────────────
...